// src/app/api/model/currency-dto-out.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CurrencyDtoOut { 
    countryCode?: string;
    id?: number;
    isoCode?: string;
    name?: string;
    originalName?: string;
    sign?: string;
}